got health working

+++ Coin_Game_Project/Assets/Scripts/Coin_Dozer/HealthManager.cs
@@ -80,7 +80,7 @@ public class HealthManager : MonoBehaviour
             //call the end of the round and reward or start next game
         }
 
-        ehpBar.fillAmount = eHP / eMaxHP;
+        ehpBar.fillAmount = (float)eHP / (float)eMaxHP;
     }
 
     #endregion
